macosx xcode 升级后编译错误及告警
fix(data_driver): 修复历史财务数据读取中的类型转换问题

在HistoryFinanceReader中添加显式类型转换以确保正确的数值比较，
避免因类型不匹配导致的数据解析错误。

feat(indicator): 防止ALIGN、MAX、MIN宏定义冲突

为ALIGN、MAX、MIN函数添加宏定义检查和取消定义，防止与系统或
其他库中的同名宏产生冲突。

fix(trade_sys): 修复System中股票代码序列化空指针访问

修改lastSuggestion方法中对空股票对象的处理逻辑，避免直接调用
空对象的方法导致程序崩溃。

chore(unit_test): 添加GCC/Clang编译器警告忽略指令

在IF指标测试文件中添加编译器诊断指令，忽略误报的缩进警告，
提高测试代码的兼容

+++ hikyuu_cpp/hikyuu/data_driver/HistoryFinanceReader.cpp
@@ -81,7 +81,7 @@ PriceList HistoryFinanceReader ::getHistoryFinanceInfo(Datetime date, const stri
         result.reserve(report_fields_count);
         price_t null_price = Null<price_t>();
         for (int i = 0; i < report_fields_count; i++) {
-            if (result_buffer[i] == 0xf8f8f8f8) {
+            if (static_cast<unsigned int>(result_buffer[i]) == 0xf8f8f8f8) {
                 result.push_back(null_price);
             } else {
                 result.push_back(result_buffer[i]);
